# Only call subscriptions when keys change