--- a/tool/get-deps.sh
+++ b/tool/get-deps.sh
@@ -2,7 +2,7 @@
 ORIG=`pwd`
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 ROOT=$DIR/../
-REPC_VERSION='v0.27.0'
+REPC_VERSION='v0.28.0'
 
 (
   cd $ROOT