Depends on repc change, https://github.com/rocicorp/repc/pull/341

Towards #63