@arv 2021-04-28
I did some tests regarding performance:

`main`

```
subscription 100 x 32.93 ops/sec ±61.2% (17 runs sampled)
subscription 500 x 12.97 ops/sec ±67.9% (8 runs sampled)
subscription 1000 x 7.55 ops/sec ±14.0% (5 runs sampled)
```

`subscribe-track`

```
subscription 100 x 52.42 ops/sec ±90.7% (27 runs sampled)
subscription 500 x 34.40 ops/sec ±279.9% (18 runs sampled)
subscription 1000 x 28.55 ops/sec ±215.4% (16 runs sampled)
```

It is clearly better but it is not O(1) as we want for a few reasons.
- We are always loading the whole map from IDB and that is O(n)
- We are running a diff over old and new value which is also O(n)