@arv 2021-04-28
I updated the test to use a constant map size and only change the number of subscriptions. The number of keys change is constant across the 3 benchmarks:

`main`

```
subscription 10 x 37.18 ops/sec ±31.6% (19 runs sampled)
subscription 100 x 25.58 ops/sec ±66.5% (13 runs sampled)
subscription 1000 x 7.06 ops/sec ±10.1% (5 runs sampled)
```

`subscribe-track`

```
subscription 10 x 37.71 ops/sec ±91.0% (19 runs sampled)
subscription 100 x 36.48 ops/sec ±152.3% (19 runs sampled)
subscription 1000 x 32.94 ops/sec ±56.8% (17 runs sampled)
```

Now we can see that things are getting closer to O(1).

I also need to restructure how subscriptions are stored in JS since there are still some O(n) (n = number of subscriptions) loops